pyproject.toml
[tool.poetry]
name = "ur3edtcourse"
version = "0.1.0"
description = "This package contains the code for the UR3e Case Study for the 'Engineering Digital Twins' course at Aarhus University."
authors = ["Andreas Kaag Thomsen <akt@ece.au.dk>"]
readme = "README.md"
packages = [
    { include = "communication" },
    { include = "startup" },
    { include = "models"},
    { include = "services"},
    { include = "utils"}
]

[tool.poetry.dependencies]
PyQt6 = "6.10.2"
python = ">=3.10,<3.12"
numpy = "<2.0"
pika = "==1.3.2"
pyhocon = "==0.3.61"
roboticstoolbox-python = "==1.1.1"
spatialmath-python = "==1.1.15"
pandas = "^2.0.0"
matplotlib = "==3.10.8"
docker = "^7.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
